Basic spring sequence setup currently springs are components

+++ Core/EngineManager.inl
@@ -388,14 +388,14 @@ namespace Gep
             Log::Critical("GetComponent() Failed, Entity: [", entity, "] does not have Component: [", GetTypeInfo<ComponentType>().PrettyName(), "]");
         }
 
-        uint64_t componentIndex = GetComponentIndex<ComponentType>();                // fast
-        Signature archetypeSignature = GetSignature(entity);                         // fast
-        glm::u64vec2 index = GetArchetypeChunkIndex(entity);                         // fast
+        uint8_t componentIndex = GetComponentIndex<ComponentType>();                 // instant?
+        Signature archetypeSignature = GetSignature(entity);                         // 1 add
+        glm::u64vec2 index = GetArchetypeChunkIndex(entity);                         // 1 add
                                                                                      
-        Archetype& archetype = mArchetypes.at(archetypeSignature);                   // slow
+        Archetype& archetype = mArchetypes.at(archetypeSignature);                   // slow: std::bitset hash
                                                                                      
-        std::byte* entityPtr = archetype.GetEntity(index.x, index.y);                // fast
-        std::byte* componentPtr = archetype.GetComponent(entityPtr, componentIndex); // fast
+        std::byte* entityPtr = archetype.GetEntity(index.x, index.y);                // 1 add 1 multiply
+        std::byte* componentPtr = archetype.GetComponent(entityPtr, componentIndex); // 2 add
 
         return *reinterpret_cast<ComponentType*>(componentPtr);
     }
